requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
amqp==5.0.2
appnope==0.1.2
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
backcall==0.2.0
billiard==3.6.3.0
blis==0.7.4
boto3==1.17.74
botocore==1.20.74
cached-property==1.5.2
catalogue==1.0.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.6
cymem==2.0.5
daphne==3.0.2
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.12
Django==3.2.3
django-allauth==0.44.0
django-cors-headers==3.7.0
django-extensions==3.1.3
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-social-oauth2==1.0.9
drf-yasg==1.20.0
geoip2==4.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
ipdb==0.13.4
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jwcrypto==0.9.1
kombu==5.0.2
MarkupSafe==1.1.1
maxminddb==2.0.3
more-itertools==8.8.0
msgpack==1.0.2
multidict==5.1.0
murmurhash==1.0.5
numpy==1.20.3
oauthlib==3.1.0
ordered-set==3.1.1
ordered-set-stubs==0.1.3
packaging==20.9
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plac==1.1.3
poetry-version==0.1.5
preshed==3.0.5
profanity-filter==1.3.3
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.7.4
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.15.100
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
scikit-learn==0.24.2
scipy==1.6.3
sentry-sdk==1.1.0
service-identity==21.1.0
six==1.15.0
sklearn==0.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
spacy==2.3.5
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.5
threadpoolctl==2.1.0
tomlkit==0.5.11
tqdm==4.60.0
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wasabi==0.8.2
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3
zope.interface==5.4.0
